+++ pkg/scheduler/framework/frameworkutils.go
@@ -7,30 +7,12 @@ package framework
 
 import (
 	"fmt"
-	"strconv"
 
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	fleetv1beta1 "go.goms.io/fleet/apis/placement/v1beta1"
 	"go.goms.io/fleet/pkg/utils"
-	"go.goms.io/fleet/pkg/utils/controller"
 )
-
-// extractNumOfClustersFromPolicySnapshot extracts the numOfClusters from the policy snapshot.
-func extractNumOfClustersFromPolicySnapshot(policy *fleetv1beta1.ClusterPolicySnapshot) (int, error) {
-	numOfClustersStr, ok := policy.Annotations[fleetv1beta1.NumberOfClustersAnnotation]
-	if !ok {
-		return 0, controller.NewUnexpectedBehaviorError(fmt.Errorf("cannot find annotation %s", fleetv1beta1.NumberOfClustersAnnotation))
-	}
-
-	// Cast the annotation to an integer; throw an error if the cast cannot be completed or the value is negative.
-	numOfClusters, err := strconv.Atoi(numOfClustersStr)
-	if err != nil || numOfClusters < 0 {
-		return 0, controller.NewUnexpectedBehaviorError(fmt.Errorf("invalid annotation %s: %s is not a valid count: %w", fleetv1beta1.NumberOfClustersAnnotation, numOfClustersStr, err))
-	}
-
-	return numOfClusters, nil
-}
 
 // extractOwnerCRPNameFromPolicySnapshot extracts the name of the owner CRP from the policy snapshot.
 func extractOwnerCRPNameFromPolicySnapshot(policy *fleetv1beta1.ClusterPolicySnapshot) (string, error) {
@@ -60,7 +42,7 @@ func classifyBindings(bindings []fleetv1beta1.ClusterResourceBinding) (active, d
 	for idx := range bindings {
 		binding := bindings[idx]
 		if binding.DeletionTimestamp != nil {
-			if controllerutil.ContainsFinalizer(&binding, utils.DispatcherFinalizer) {
+			if controllerutil.ContainsFinalizer(&binding, fleetv1beta1.DispatcherFinalizer) {
 				deletedWithDispatcherFinalizer = append(deletedWithDispatcherFinalizer, &binding)
 			} else {
 				deletedWithoutDispatcherFinalizer = append(deletedWithoutDispatcherFinalizer, &binding)
